india-price-intelligence-pipeline/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-bigquery==3.13.0
pandas==2.1.0
python-dotenv==1.0.0
dbt-bigquery==1.7.0
apache-airflow==2.8.0
faker==22.0.0
selenium==4.17.2
